Exit early in some situations when the control has been removed (#8)
There are some cases where the control might get removed while a function is still running.

For example, the render function fires off the event `Constants.events.DELETE`, and the event handling might cause the control to get removed. If this is the case, then trying to fire a render event later on in the render function would fail. Instead, when the control did get removed, we should clean up and exit early.

Similarly, after calling `trash` on the mode, the control might have gotten removed. So check for that case and exit early if so.